Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -6,10 +6,10 @@ int main() {
     int o,j;
     scanf("%d %d %c",&a,&b,&c);
     if( c == '+') {
-        int u = a+b;
-        printf("%f",u);
+        int u =(int)a+b;
+        printf("%d",u);
     } else if (c == '-') {
-        int u = a-b;
+        int u =(int)a-b;
         printf("%d",u);
     } else if(c == '*') {
         float e = a*b;
